pyproject.toml
[project]
name = "gui-agent"
version = "0.1.0"
description = "GUI Agent prototype using Google ADK with Playwright MCP tools"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Ryan", email = "rwesslen@gmail.com" }
]

dependencies = [
    # ADK Framework
    "google-adk>=0.5.0",
    "google-genai>=1.0.0",

    # Observability
    "arize-phoenix>=8.0.0",
    "openinference-instrumentation-google-genai>=0.1.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",

    # Configuration
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # Async support
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

mock-server = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
gui-agent = "gui_agent_v1.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gui_agent_v1"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
    "evalset: marks tests using ADK evaluation format",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
